use fragment pagination with top level query
